split task tests
