+++ tests/tasks/conftest.py
@@ -1,0 +1,77 @@
+import pytest
+import pandas as pd
+import anndata as ad
+from czbenchmarks.tasks.types import CellRepresentation
+from czbenchmarks.datasets.types import Organism
+from tests.utils import create_dummy_anndata
+
+
+@pytest.fixture
+def dummy_anndata():
+    n_cells: int = 500
+    n_genes: int = 200
+    organism: Organism = Organism.HUMAN
+    obs_columns: list[str] = ["cell_type", "batch"]
+    var_columns: list[str] = ["feature_name"]
+    anndata: ad.AnnData = create_dummy_anndata(
+        n_cells=n_cells,
+        n_genes=n_genes,
+        organism=organism,
+        obs_columns=obs_columns,
+        var_columns=var_columns,
+    )
+
+    expression_matrix: CellRepresentation = anndata.X.copy()
+    obs: pd.DataFrame = anndata.obs.copy()
+    var: pd.DataFrame = anndata.var.copy()
+
+    # TODO perform PCA on expression matrix to get true embedding
+    embedding_matrix: CellRepresentation = expression_matrix.toarray()
+
+    return {
+        "anndata": anndata,
+        "expression_matrix": expression_matrix,
+        "obs": obs,
+        "var": var,
+        "embedding_matrix": embedding_matrix,
+    }
+
+
+@pytest.fixture
+def expression_matrix(dummy_anndata):
+    return dummy_anndata["expression_matrix"]
+
+
+@pytest.fixture
+def embedding_matrix(dummy_anndata):
+    return dummy_anndata["embedding_matrix"]
+
+
+@pytest.fixture
+def obs(dummy_anndata):
+    return dummy_anndata["obs"]
+
+
+@pytest.fixture
+def var(dummy_anndata):
+    return dummy_anndata["var"]
+
+
+@pytest.fixture
+def fixture_data(request):
+    # Enables lazy generation of fixture data so fixtures can be used as
+    # parameters
+    valid_fixture_names = ["expression_matrix", "embedding_matrix", "obs", "var"]
+    fixture_name, other_data = request.param
+    if isinstance(fixture_name, str):
+        fixture_data = (
+            request.getfixturevalue(fixture_name)
+            if fixture_name in valid_fixture_names
+            else fixture_name
+        )
+    else:
+        fixture_data = [
+            request.getfixturevalue(f) if f in valid_fixture_names else f
+            for f in fixture_name
+        ]
+    return fixture_data, other_data
