+++ tests/tasks/test_cross_species_integration.py
@@ -1,0 +1,37 @@
+import pytest
+from czbenchmarks.tasks.single_cell import (
+    CrossSpeciesIntegrationTask,
+    CrossSpeciesIntegrationTaskInput,
+)
+from czbenchmarks.datasets.types import Organism
+from czbenchmarks.metrics.types import MetricResult
+
+
+def test_cross_species_task(embedding_matrix, obs):
+    """Test that CrossSpeciesIntegrationTask executes without errors."""
+    task = CrossSpeciesIntegrationTask()
+    embedding_list = [embedding_matrix, embedding_matrix]
+    labels = obs["cell_type"]
+    labels_list = [labels, labels]
+    organism_list = [Organism.HUMAN, Organism.MOUSE]
+    task_input = CrossSpeciesIntegrationTaskInput(
+        labels=labels_list, organism_list=organism_list
+    )
+
+    try:
+        # Test regular task execution
+        results = task.run(
+            cell_representation=embedding_list,
+            task_input=task_input,
+        )
+
+        # Verify results structure
+        assert isinstance(results, list)
+        assert all(isinstance(r, MetricResult) for r in results)
+
+        # Test that baseline raises NotImplementedError
+        with pytest.raises(NotImplementedError):
+            task.compute_baseline()
+
+    except Exception as e:
+        pytest.fail(f"CrossSpeciesIntegrationTask failed unexpectedly: {e}")
