+++ tests/tasks/test_general_tasks.py
@@ -1,0 +1,144 @@
+import pytest
+from czbenchmarks.tasks import (
+    ClusteringTask,
+    ClusteringTaskInput,
+    EmbeddingTask,
+    EmbeddingTaskInput,
+    BatchIntegrationTask,
+    BatchIntegrationTaskInput,
+    MetadataLabelPredictionTask,
+    MetadataLabelPredictionTaskInput,
+)
+from czbenchmarks.metrics.types import MetricResult
+from tests.utils import DummyTask, DummyTaskInput
+
+
+@pytest.mark.parametrize(
+    "fixture_data",
+    [
+        ("expression_matrix", False),
+        (["expression_matrix", "expression_matrix"], True),
+    ],
+    indirect=True,
+)
+def test_embedding_valid_input_output(fixture_data):
+    """Test that embedding is accepted and List[MetricResult] is returned."""
+    embedding, requires_multiple_datasets = fixture_data
+    task = DummyTask(requires_multiple_datasets=requires_multiple_datasets)
+    results = task.run(
+        cell_representation=embedding,
+        task_input=DummyTaskInput(),
+    )
+
+    assert isinstance(results, list)
+    assert all(isinstance(r, MetricResult) for r in results)
+
+
+@pytest.mark.parametrize(
+    "fixture_data",
+    [
+        (
+            "abcd",
+            [False, "This task requires a single cell representation for input"],
+        ),
+        (
+            ["embedding_matrix"],
+            [False, "This task requires a single cell representation for input"],
+        ),
+        (
+            ["embedding_matrix", "embedding_matrix"],
+            [False, "This task requires a single cell representation for input"],
+        ),
+        (
+            "embedding_matrix",
+            [True, "This task requires a list of cell representations"],
+        ),
+        (
+            ["abcd", "embedding_matrix"],
+            [True, "This task requires a list of cell representations"],
+        ),
+        (
+            ["embedding_matrix"],
+            [
+                True,
+                "This task requires a list of cell representations",
+            ],
+        ),
+    ],
+    indirect=True,
+)
+def test_embedding_invalid_input(fixture_data):
+    """Test ValueError for mismatch with requires_multiple_datasets."""
+    embedding_list, (requires_multiple_datasets, error_message) = fixture_data
+    task = DummyTask(requires_multiple_datasets=requires_multiple_datasets)
+    with pytest.raises(ValueError, match=error_message):
+        task.run(
+            cell_representation=embedding_list,
+            task_input=DummyTaskInput(),
+        )
+
+
+@pytest.mark.parametrize(
+    "task_class,task_input_builder",
+    [
+        (
+            ClusteringTask,
+            lambda obs: ClusteringTaskInput(obs=obs, input_labels=obs["cell_type"]),
+        ),
+        (
+            EmbeddingTask,
+            lambda obs: EmbeddingTaskInput(input_labels=obs["cell_type"]),
+        ),
+        (
+            BatchIntegrationTask,
+            lambda obs: BatchIntegrationTaskInput(
+                labels=obs["cell_type"], batch_labels=obs["batch"]
+            ),
+        ),
+        (
+            MetadataLabelPredictionTask,
+            lambda obs: MetadataLabelPredictionTaskInput(labels=obs["cell_type"]),
+        ),
+    ],
+)
+def test_task_execution(
+    task_class,
+    task_input_builder,
+    embedding_matrix,
+    expression_matrix,
+    obs,
+):
+    """Test that each task executes without errors on compatible data."""
+
+    task_input = task_input_builder(obs)
+
+    task = task_class()
+
+    try:
+        # Test regular task execution
+        results = task.run(
+            cell_representation=embedding_matrix,
+            task_input=task_input,
+        )
+        assert isinstance(results, list)
+        assert all(isinstance(r, MetricResult) for r in results)
+
+        # Test baseline execution if implemented
+        try:
+            n_pcs = min(50, expression_matrix.shape[1] - 1)
+            baseline_embedding = task.compute_baseline(expression_matrix, n_pcs=n_pcs)
+            if hasattr(task_input, "var"):
+                task_input.var = task_input.var.iloc[:n_pcs]
+
+            baseline_results = task.run(
+                cell_representation=baseline_embedding,
+                task_input=task_input,
+            )
+            assert isinstance(baseline_results, list)
+            assert all(isinstance(r, MetricResult) for r in baseline_results)
+        except NotImplementedError:
+            # Some tasks may not implement compute_baseline
+            pass
+
+    except Exception as e:
+        pytest.fail(f"Task {task_class.__name__} failed unexpectedly: {e}")
